refactor: Add explicit integer casts for lift length and elevation gain in cost calculation and include the `System.Collections.Generic` namespace.

+++ Assets/Scripts/Core/LiftSystem.cs
@@ -107,8 +107,8 @@ namespace SkiResortTycoon.Core
         public int CalculateCost(LiftData lift)
         {
             int cost = BaseCost;
-            cost += lift.Length * CostPerTile;
-            cost += lift.ElevationGain * CostPerHeightUnit;
+            cost += (int)lift.Length * CostPerTile;
+            cost += (int)lift.ElevationGain * CostPerHeightUnit;
             
             lift.BuildCost = cost;
             return cost;
